requirements.txt
google
pytz
pyTelegramBotAPI
selenium
fake_useragent
python-dotenv